v 1.5: New import/export feature, faster launch
The import works with an embedded HTTP server that allows you to upload
report files that you've downloaded elsewhere, for example via AppViz
on your Mac or directly from itts.

Export works similarly; you get a list of your reports in AppSales
that you can simply download to your computer from the browser.
This only works for reports that you downloaded starting with this
version, because earlier versions did not store the original report
files.

To speed up launching the app, some report data is now stored in a cache
file. Generating this file for the first time after upgrading from
an older version may take a while.